// client/model/f5RuntimePoolMember.ts
/**
 * Draas
 * Draas OAS Documentation
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { F5RuntimeFqdnFqdn } from './f5RuntimeFqdnFqdn';


export interface F5RuntimePoolMember { 
    /**
     * Member name
     */
    readonly name: string;
    /**
     * Member address
     */
    readonly address: string;
    /**
     * Ephemeral state
     */
    readonly ephemeral: string;
    /**
     * FQDN configuration
     */
    readonly fqdn: F5RuntimeFqdnFqdn;
    /**
     * Priority group
     */
    readonly priorityGroup: number;
    /**
     * Ratio
     */
    readonly ratio: number;
    /**
     * State
     */
    readonly state: string;
    /**
     * Full path
     */
    readonly fullPath: string;
}