// client/model/f5RuntimeVirtualServer.ts
/**
 * Draas
 * Draas OAS Documentation
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { F5RuntimePoliciesReference } from './f5RuntimePoliciesReference';
import { F5RuntimeProfilesReference } from './f5RuntimeProfilesReference';
import { F5RuntimePoolReference } from './f5RuntimePoolReference';
import { F5RuntimeCertsReference } from './f5RuntimeCertsReference';
import { F5RuntimeNestedStats } from './f5RuntimeNestedStats';


export interface F5RuntimeVirtualServer { 
    /**
     * Virtual server name
     */
    readonly name: string;
    /**
     * Description of the virtual server
     */
    readonly description: string;
    /**
     * Partition name
     */
    readonly partition: string;
    /**
     * Whether the virtual server is enabled
     */
    readonly enabled: boolean;
    /**
     * Address status
     */
    readonly addressStatus: string;
    /**
     * Connection limit
     */
    readonly connectionLimit: number;
    /**
     * IP protocol
     */
    readonly ipProtocol: string;
    /**
     * Destination address
     */
    readonly destination: string;
    /**
     * Subnet mask
     */
    readonly mask: string;
    /**
     * Reference to the pool associated with the virtual server
     */
    readonly poolReference: F5RuntimePoolReference;
    /**
     * Reference to profiles associated with the virtual server
     */
    readonly profilesReference: F5RuntimeProfilesReference;
    /**
     * Reference to policies associated with the virtual server
     */
    readonly policiesReference: F5RuntimePoliciesReference;
    /**
     * Reference to certificates associated with the virtual server
     */
    readonly certsReference: Array<F5RuntimeCertsReference>;
    /**
     * Statistics for the virtual server
     */
    readonly stats: F5RuntimeNestedStats;
}